refactor(cvaos): give each data table a single collection name
Issue: Every data submodule published its rows twice, once as an immutable
rows tuple and again as a mutable list copy under a descriptive name. The two
held the same records, so callers picked between them arbitrarily and some
reached past the data facade to do it.

Solution: The rows tuple is now the only collection each submodule defines,
and the data package re-exports it under the descriptive name as the same
object. Callers that imported a submodule directly now go through the facade.
