+++ worlds/cvaos/data/__init__.py
@@ -1,10 +1,22 @@
+"""
+Parsed Castlevania: Aria of Sorrow game data.
+
+Each subpackage loads its CSVs once at import and exposes the result as an immutable
+``rows`` tuple plus whatever lookup dicts it needs. This module is the single import
+point for the rest of the world, so it re-exports each ``rows`` tuple under a name that
+says which table it came from. The alias is the same object, not a copy.
+
+Route solving over this data lives in ``worlds/cvaos/tools/routing``, which may import
+from here but is never imported by it.
+"""
+
 from __future__ import annotations
 
-from .entrance_info import EntranceInfo, entrance_info_collection
-from .item_balancing import DesirabilityInfo, desirability_collection
-from .item_info import ItemInfo, item_info_collection
-from .pickup_info import PickupInfo, pickup_info_collection
-from .room_info import RoomInfo, room_info_collection
+from .entrance_info import EntranceInfo, rows as entrance_info_collection
+from .item_balancing import DesirabilityInfo, rows as desirability_collection
+from .item_info import ItemInfo, rows as item_info_collection
+from .pickup_info import PickupInfo, rows as pickup_info_collection
+from .room_info import RoomInfo, rows as room_info_collection
 from .routing_info import (
     AbilityCombo,
     EntranceToPickupRegionInfo,
@@ -15,12 +27,12 @@ from .routing_info import (
     by_enemy_number_for_enemy_regions,
     by_from_entrance_for_transdoor,
     enemy_meta_by_number,
-    entrance_to_entrance_info_collection,
-    entrance_to_enemy_region_info_collection,
-    entrance_to_pickup_region_info_collection,
+    enemy_region_rows as entrance_to_enemy_region_info_collection,
     lookup_pickup_region_requirement,
+    pickup_region_rows as entrance_to_pickup_region_info_collection,
     resolve_enemy_number,
-    transdoor_connection_collection,
+    rows as entrance_to_entrance_info_collection,
+    transdoor_connection_rows as transdoor_connection_collection,
 )
 
 __all__ = [
