+++ worlds/cvaos/tools/routing/entrances_to_items.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Dict, Iterable, List, Mapping, Optional, Sequence, Set, Tuple
 
-from ...data.routing_info import (
+from ...data import (
     EntranceToPickupRegionInfo,
     entrance_to_entrance_info_collection,
     entrance_to_pickup_region_info_collection,
